requirements.txt
django-cms<=3.0.7
django-simple-captcha==0.4.9
